[DNM] opt: add exploration rule for join associativity
This commit adds an exploration rule for join associativity. The new rule
converts an expression like:
   A JOIN (B JOIN C ON B.x = C.x) ON A.y = B.y
to the logically equivalent expression:
   (A JOIN B ON A.y = B.y) JOIN C ON B.x = C.x
The rule applies as long as the outer predicate is not True (we want to avoid
pushing cross joins down), and the outer predicate only references one of
the two inner relations.

I don't think we should merge this commit until we have join hints and/or
automated collection of table stats (and possibly more aggressive pruning
of the search space), but just wanted share how easy it is to add the
capability for join reordering.

Release note: None