[DNM] opt: add encode test directive to OptTester
This commit adds an "encode" test directive to the OptTester. When the
directive is used, the OptTester builds an expression tree from a SQL
query, fully optimizes it using the memo, and then outputs the lowest
cost tree in a format that can be understood by a neural network.

This commit also adds a testing flag called "count" which can be used
to indicate the number of times to perform optimization. It can be used
in combination with the perturb-cost flag to display several alternate
output plans for the same query. It currently only works with the encode
command. Plans are de-duplicated and printed in order of increasing cost.

Release note: None