sql,opt: export pg_catalog table schemas for use by optimizer
This commit moves the schema definitions of all tables in pg_catalog to
a new file in sql/vtable called pg_catalog.go and exports them so they
can be used by the optimizer and elsewhere. This is similar to what was
already done for the information_schema tables.

Release note: None